requirements.txt
# AI狼人杀游戏依赖
# 项目基本运行环境

# API客户端，用于AI玩家
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.2.0
requests>=2.28.0

# 数据处理和分析
pandas>=1.3.0

# 日期时间处理工具
python-dateutil>=2.8.2

# 文本处理相关
regex>=2022.1.18

# 测试框架
pytest>=7.0.0

# 开发工具
flake8>=4.0.1
black>=22.1.0

# Web服务相关(如果有Web界面)
flask>=2.0.0
jinja2>=3.0.0

# 错误追踪和日志
sentry-sdk>=1.14.0

# 系统环境变量管理
python-dotenv>=0.20.0 